Added FIJI Blurring. Only works colour -> b/w at the moment.
I already dislike FIJI due to the sevre number of jars required to run it, which, in my experience, leads to classpath issues in Java.

The API is difficult to use because of this; not knowing which package or jar a given class comes from. OpenCV has the edge personally.